requirements.txt
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
aniso8601==9.0.1
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
diffusers==0.19.3
easydict==1.10
enum-compat==0.0.3
exceptiongroup==1.1.2
fastapi==0.100.1
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
gdown==4.7.1
gradio==3.39.0
gradio_client==0.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
networkx==3.1
numpy==1.25.1
nvgpu==0.10.0
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rpds-py==0.9.2
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tabulate==0.9.0
termcolor==2.3.0
tifffile==2023.7.18
tokenizers==0.13.3
toolz==0.12.0
torch-model-archiver==0.8.1
torchserve==0.8.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
yarl==1.9.2
zipp==3.16.2
